requirements.txt
cleverbot
steamodd